requirement.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
matplotlib
scikit-learn